### 2020-01-06 #1993 @whedon
[PRE REVIEW]: CMasher: Scientific colormaps for making accessible, informative and 'cmashing' plots
**Submitting author:** @1313e (<a href="http://orcid.org/0000-0002-1559-9832">Ellert van der Velden</a>)
**Repository:** <a href="https://github.com/1313e/CMasher" target ="_blank">https://github.com/1313e/CMasher</a>
**Version:** v1.2.0
