### 2020-01-06 #1993 @danielskatz
👋 @mbobra - would you be willing to edit this submission for JOSS? (Note that one reviewer has already volunteered)
