### 2020-01-07 40f6abf
Few updates to the JOSS paper.
