+++ joss_paper/paper.md
@@ -21,33 +21,40 @@ bibliography: paper.bib
 
 # Introduction
 
-The use of colorizations in the visualization of scientific results is a common sight nowadays.
-It allows for more (complex) data to be plotted in the same figure without having to resort to difficult-to-read 3D plots or subplots; online material; or interactive applications.
-However, an often underappreciated aspect of the choice of colormap, is how colorization affects the way the visualized data is interpreted.
+The use of colors in the visualization of scientific results is a common sight nowadays.
+Color allows for more (complex) data to be plotted in the same figure without resorting to difficult-to-interpret 3D plots or subplots; online material; or interactive applications.
+However, an often underappreciated aspect of data visualization, is how color affects the way the visualized data is interpreted, making it crucial to pick the correct colormap.
+In order to help with picking a scientific colormap, I introduce the *CMasher* package.
 
 
 # Background summary
 
-A good scientific colormap is often described/characterized as _perceptually uniform sequential_ [@Rogowitz96; @Sharpe99], which means that it is perceived as uniformly changing in lightness and saturation, mostly at the same hue.
+A good scientific colormap is often described/characterized as _perceptually uniform sequential_, which means that the colormap is perceived as uniformly changing in lightness and saturation, mostly at the same hue [@Rogowitz96; @Sharpe99].
 This allows for the data values of a plot to be interpreted correctly by the viewer without giving false information.
-It also often allows for a plot to be converted properly to grey-scale without losing information.
-Basically, it should be possible to infer how every color relates in terms of value to every other color, by simply looking at it (so, a legend or colorbar should NOT be necessary for this task).
+Such a colormap also often allows for a plot to be converted properly to grey-scale without losing information.
+A perceptually uniform sequential colormap allows us to properly infer the relative order of the represented numerical values, without requiring a legend or colorbar.
 
 Although there are many works out there that describe the optimal way to do this [@Rogowitz96; @Sharpe99; @Kindlmann02; @Birch12; @Brychtova16; @Szafir18] and there are tools readily available to test the performance of a colormap [@cmaputil; @colorspacious; @viscm], bad/misleading colormaps are still very commonly used.
-The main problem usually is that humans do not perceive every color equally (e.g., small variations in the color green are not perceived as it is such a natural color, while small variations in the colors red and blue are perceived).
-A good example of a colormap that illustrates this, is the commonly used *jet* colormap:
+The main issue usually is that humans do not perceive every color equally (e.g., small variations in the color green are not perceived as green is a common natural color, while small variations in the colors red and blue are perceived).
+Here, we use the *jet* colormap to illustrate this issue:
 
-![Output of the ``viscm`` package showing the statistics and performance of the *jet* colormap. The various different plots show how the colormap changes in perceived color and lightness, as well as how well the colormap converts to different types of color-vision deficiency (CVD; color blindness) and grey-scale. In case of a perceptually uniform sequential colormap, the two derivative plots should show a straight horizontal line; the colorspace diagram should be smooth; and the lines in the right-bottom corner plots should be visible up to the same depth across the entire colormap.](https://raw.githubusercontent.com/1313e/CMasher/master/docs/source/user/images/jet_viscm.png)
[…]
