+++ docs/source/user/background.rst
@@ -1,13 +1,13 @@
 Background
 ==========
-A good scientific colormap is often described/characterized as *perceptually uniform sequential*, which means that it is perceived as uniformly changing in lightness and saturation, mostly at the same hue.
-This allows for the data values of a plot to be interpreted correctly by the viewer without giving false information (a great example of this can be found `here <https://mycarta.wordpress.com/2012/10/14/the-rainbow-is-deadlong-live-the-rainbow-part-4-cie-lab-heated-body/>`_).
-It also often allows for a plot to be converted properly to grey-scale without losing information.
-Basically, it should be possible to infer how every color relates in terms of value to every other color, by simply looking at it (so, a legend or colorbar should NOT be necessary for this task).
+A good scientific colormap is often described/characterized as *perceptually uniform sequential*, which means that the colormap is perceived as uniformly changing in lightness and saturation, mostly at the same hue.
+This allows for the data values of a plot to be interpreted correctly by the viewer without giving false information (a great explanation of this can be found `here <https://mycarta.wordpress.com/2012/10/14/the-rainbow-is-deadlong-live-the-rainbow-part-4-cie-lab-heated-body/>`_).
+Such a colormap also often allows for a plot to be converted properly to grey-scale without losing information.
+A perceptually uniform sequential colormap allows us to properly infer the relative order of the represented numerical values, without requiring a legend or colorbar.
 
 Although this may sound easy enough, there are many colormaps out there that do this incorrectly.
-The main problem usually is that humans do not perceive every color equally (e.g., small variations in the color green are not perceived as it is such a natural color, while small variations in the colors red and blue are perceived).
-A good example of a colormap that illustrates this, is the commonly used *jet* colormap:
+The main issue usually is that humans do not perceive every color equally (e.g., small variations in the color green are not perceived as green is a common natural color, while small variations in the colors red and blue are perceived).
+An example of a colormap that illustrates this issue, is the *jet* colormap:
 
 .. figure:: images/jet_viscm.png
     :alt: Statistics of the *jet* colormap.
@@ -17,16 +17,22 @@ A good example of a colormap that illustrates this, is the commonly used *jet* c
 
     Output of the *viscm* package showing the statistics and performance of the *jet* colormap.
     The various different plots show how the colormap changes in perceived color and lightness, as well as how well the colormap converts to different types of color-vision deficiency (CVD; color blindness) and grey-scale.
-    In case of a perceptually uniform sequential colormap, the two derivative plots should show a straight horizontal line; the colorspace diagram should be smooth; and the lines in the right-bottom corner plots should be visible up to the same depth across the entire colormap.
+    In case of a perceptually uniform sequential colormap, the two derivative plots should show a straight horizontal line; the colorspace diagram should be smooth; and the lines in the bottom-right corner plots should be visible up to the same depth across the entire colormap.
 
-In :numref:`jet_viscm`, one can view the statistics output of the *jet* colormap, made with the `viscm`_ package.
-This colormap shows the spectrum of visible light, which trivially increases linearly in wavelength.
-However, we can clearly see that this introduces multiple problems, as the color green for example is perceived as the brightest of the visible colors due to its very natural occurance, and the colormap is absolutely not CVD-friendly.
-This is an example of a colormap where it would be necessary to have a colorbar/legend, and it is overall a bad colormap to use.
+In :numref:`jet_viscm`, one can view the performance output of the *jet* colormap, made with the `viscm`_ package.
+For perceptually uniform sequential colormaps, the two derivative plots in the top-left should show a straight horizontal line, indicating that the colormap changes uniformly in both perceived saturation and lightness.
+Consequently, the colorspace diagram in the bottom-left should be smooth.
[…]
