### 2020-01-07 #1993 @arfon
> 👋 @arfon - can you see the problem here?

Sorry, this was an unintended side effect of upgrading Pandoc last night. It's now fixed.
