### 2020-01-08 #2004 @1313e
@mbobra Can you remind me, after the review process has finished, to do something about the placement of Figure 2 in the paper?
I will probably remember it, but just to be sure.
