### 2020-01-10 309c514
Few very small wording changes.
