### 2020-01-13 6d7415e
Fixed a bug where providing the name of a reversed colormap to create_cmap_overview() would not work properly.
