### 2020-01-14 69b008b
The create_cmap_overview()-function can now sort colormaps based on their lightness profile. Redrawn the viscm-outputs for all diverging colormaps.
