+++ cmasher/utils.py
@@ -74,7 +74,7 @@ def _get_cm_type(cmap):
     diff_L1 = np.diff(L[central_i[1]:])
 
     # Obtain perceptual differences of last two and first two values
-    lab_red = np.concatenate([lab[-2:], lab[:2]], axis=0)
+    lab_red = lab[[-2, -1, 0, 1]]
     deltas = np.sqrt(np.sum(np.diff(lab_red, axis=0)**2, axis=-1))
 
     # Check the statistics of cmap and determine the colormap type
@@ -113,9 +113,73 @@ def _get_cm_type(cmap):
         return('misc')
 
 
+# Define function for obtaining the sorting order for lightness ranking
+def _get_cmap_lightness_rank(cmap):
+    """
+    Returns a tuple of objects used for sorting the provided `cmap` based
+    on its lightness profile.
+
+    Parameters
+    ----------
+    cmap : str or :obj:`~matplotlib.colors.Colormap` object
+        The registered name of the colormap in *MPL* or its corresponding
+        :obj:`~matplotlib.colors.Colormap` object.
+
+    Returns
+    -------
+    L_start : float
+        The starting lightness value of `cmap`.
+    L_rmse : float
+        The RMSE of the lightness profile of `cmap`.
+        For diverging colormaps, this is the max RMSE of either half.
+    L_rng : float
+        The lightness range (L_max-L_min) of `cmap`.
+    name : str
+        The name of `cmap`.
+
+    """
+
+    # Obtain the colormap
+    cmap = mplcm.get_cmap(cmap)
+
+    # Get array of all values for which a colormap value is requested
+    x = np.linspace(0, 1, cmap.N)
+
+    # Get RGB values for colormap
+    rgb = cmap(x)[:, :3]
+
+    # Get lightness values of colormap
+    lab = cspace_converter("sRGB1", "CAM02-UCS")(rgb)
+    L = lab[:, 0]
+
+    # Determine the deltas of the lightness profile
+    deltas = np.diff(L)
+    derivs = (cmap.N-1)*deltas
+
+    # Determine the RMSE of the lightness profile
+    if _get_cm_type(cmap) in ('diverging', 'cyclic'):
+        # If cmap is diverging, calculate RMSE of both halves
+        central_i = [int(np.floor(cmap.N/2)), int(np.ceil(cmap.N/2))]
+        L_rmse = np.max([np.around(np.std(derivs[:central_i[0]]), 1),
+                         np.around(np.std(derivs[central_i[1]:]), 1)])
+    else:
+        # Else, take RMSE of entire lightness profile
+        L_rmse = np.around(np.std(derivs), 1)
+
+    # Determine start and range
+    L_start = np.around(L[0], 1)
+    L_min = np.around(np.min(L), 1)
+    L_max = np.around(np.max(L), 1)
+    L_rng = np.around(np.abs(L_max-L_min), 1)
+
+    # Return contributions to the rank
+    return(L_start, L_rmse, L_rng, cmap.name)
+
+
 # %% FUNCTIONS
 # This function creates an overview plot of all colormaps specified
-def create_cmap_overview(cmaps=None, savefig=None, use_types=True):
+def create_cmap_overview(cmaps=None, savefig=None, use_types=True,
+                         sort='alphabetical'):
     """
     Creates an overview plot containing all colormaps defined in the provided
     `cmaps`.
@@ -134,6 +198,12 @@ def create_cmap_overview(cmaps=None, savefig=None, use_types=True):
         Whether all colormaps in `cmaps` should be categorized into their
         colormap types (sequential; diverging; cyclic; qualitative; misc).
         If `cmaps` is a dict, this value is ignored.
[…]
