### 2020-01-25 eeb28a7
Added the 'lavender' and 'nuclear' sequential colormaps.
