+++ cmasher/colormaps/nuclear/nuclear.py
@@ -1,0 +1,262 @@
+from matplotlib.colors import ListedColormap
+
+cm_type = "linear"
+
+cm_data = [[0.00000000e+00, 0.00000000e+00, 0.00000000e+00],
+           [2.18978444e-04, 1.77540646e-04, 2.55612013e-04],
+           [7.68124580e-04, 6.08776467e-04, 9.25389883e-04],
+           [1.60743304e-03, 1.24801543e-03, 1.99558151e-03],
+           [2.72048166e-03, 2.07363753e-03, 3.47485305e-03],
+           [4.09715583e-03, 3.07139249e-03, 5.37786044e-03],
+           [5.73018628e-03, 4.23073344e-03, 7.72233034e-03],
+           [7.61380934e-03, 5.54335658e-03, 1.05279395e-02],
+           [9.74316966e-03, 7.00241938e-03, 1.38160755e-02],
+           [1.21139195e-02, 8.60215564e-03, 1.76093743e-02],
+           [1.47220226e-02, 1.03375904e-02, 2.19317037e-02],
+           [1.75636340e-02, 1.22043364e-02, 2.68082763e-02],
+           [2.06349585e-02, 1.41985119e-02, 3.22653727e-02],
+           [2.39321803e-02, 1.63166491e-02, 3.83303458e-02],
+           [2.74514250e-02, 1.85556037e-02, 4.48143024e-02],
+           [3.11887041e-02, 2.09125099e-02, 5.13195203e-02],
+           [3.51398654e-02, 2.33847472e-02, 5.78545958e-02],
+           [3.93005180e-02, 2.59699639e-02, 6.44243769e-02],
+           [4.35416128e-02, 2.86659924e-02, 7.10332953e-02],
+           [4.77103734e-02, 3.14708616e-02, 7.76853131e-02],
+           [5.18184827e-02, 3.43827896e-02, 8.43839988e-02],
+           [5.58674191e-02, 3.74001823e-02, 9.11325837e-02],
+           [5.98582631e-02, 4.05203677e-02, 9.79340035e-02],
+           [6.37917313e-02, 4.36167095e-02, 1.04790928e-01],
+           [6.76681997e-02, 4.66743313e-02, 1.11705782e-01],
+           [7.14877210e-02, 4.96957395e-02, 1.18680754e-01],
+           [7.52500517e-02, 5.26832389e-02, 1.25718267e-01],
+           [7.89545993e-02, 5.56392251e-02, 1.32819807e-01],
+           [8.26004715e-02, 5.85660208e-02, 1.39986815e-01],
+           [8.61864649e-02, 6.14658124e-02, 1.47221131e-01],
+           [8.97110198e-02, 6.43409488e-02, 1.54523749e-01],
+           [9.31722230e-02, 6.71938134e-02, 1.61895543e-01],
+           [9.65677748e-02, 7.00268571e-02, 1.69337261e-01],
+           [9.98949472e-02, 7.28426366e-02, 1.76849496e-01],
+           [1.03150569e-01, 7.56440312e-02, 1.84431780e-01],
+           [1.06330947e-01, 7.84340173e-02, 1.92083865e-01],
+           [1.09431808e-01, 8.12158616e-02, 1.99804968e-01],
+           [1.12448240e-01, 8.39931890e-02, 2.07593650e-01],
+           [1.15374609e-01, 8.67700585e-02, 2.15447655e-01],
+           [1.18204478e-01, 8.95510506e-02, 2.23363720e-01],
+           [1.20930510e-01, 9.23413649e-02, 2.31337335e-01],
+           [1.23544151e-01, 9.51466855e-02, 2.39363937e-01],
[…]
