+++ cmasher/__version__.py
@@ -10,4 +10,4 @@ Stores the different versions of the *CMasher* package.
 
 # %% VERSIONS
 # Default/Latest/Current version
-__version__ = '1.2.0'
+__version__ = '1.2.1'
