+++ cmasher/colormaps/lavender/lavender.py
@@ -1,0 +1,262 @@
+from matplotlib.colors import ListedColormap
+
+cm_type = "linear"
+
+cm_data = [[0.00000000e+00, 0.00000000e+00, 0.00000000e+00],
+           [2.26276605e-04, 1.27312718e-04, 1.80411323e-04],
+           [8.13645546e-04, 4.28891491e-04, 6.39834919e-04],
+           [1.74260184e-03, 8.63685259e-04, 1.35487268e-03],
+           [3.01237471e-03, 1.41020371e-03, 2.32124316e-03],
+           [4.62644578e-03, 2.05319508e-03, 3.54022725e-03],
+           [6.59022229e-03, 2.78050385e-03, 5.01612876e-03],
+           [8.90954870e-03, 3.58208100e-03, 6.75497888e-03],
+           [1.15908560e-02, 4.44910184e-03, 8.76435018e-03],
+           [1.46413908e-02, 5.37339649e-03, 1.10533959e-02],
+           [1.80675978e-02, 6.34788166e-03, 1.36318649e-02],
+           [2.18766340e-02, 7.36569349e-03, 1.65109504e-02],
+           [2.60757165e-02, 8.42034129e-03, 1.97029401e-02],
+           [3.06714488e-02, 9.50591564e-03, 2.32207803e-02],
+           [3.56714051e-02, 1.06162932e-02, 2.70791788e-02],
+           [4.10685016e-02, 1.17460710e-02, 3.12931556e-02],
+           [4.65285211e-02, 1.28895898e-02, 3.58795272e-02],
+           [5.19623992e-02, 1.40414219e-02, 4.08481551e-02],
+           [5.73725657e-02, 1.51965674e-02, 4.59230286e-02],
+           [6.27621826e-02, 1.63494237e-02, 5.10293953e-02],
+           [6.81330369e-02, 1.74950091e-02, 5.61721676e-02],
+           [7.34866739e-02, 1.86283781e-02, 6.13560706e-02],
+           [7.88246360e-02, 1.97444393e-02, 6.65859302e-02],
+           [8.41481990e-02, 2.08381282e-02, 7.18664189e-02],
+           [8.94585856e-02, 2.19042363e-02, 7.72023166e-02],
+           [9.47563555e-02, 2.29379567e-02, 8.25977765e-02],
+           [1.00042052e-01, 2.39343733e-02, 8.80570739e-02],
+           [1.05316089e-01, 2.48885467e-02, 9.35844918e-02],
+           [1.10578791e-01, 2.57954689e-02, 9.91843886e-02],
+           [1.15830405e-01, 2.66500344e-02, 1.04861240e-01],
+           [1.21070636e-01, 2.74476067e-02, 1.10618951e-01],
+           [1.26299179e-01, 2.81834467e-02, 1.16461601e-01],
+           [1.31515560e-01, 2.88529100e-02, 1.22393213e-01],
+           [1.36719884e-01, 2.94503632e-02, 1.28419043e-01],
+           [1.41910804e-01, 2.99720524e-02, 1.34542190e-01],
+           [1.47087679e-01, 3.04131133e-02, 1.40767177e-01],
+           [1.52249518e-01, 3.07690478e-02, 1.47098213e-01],
+           [1.57395119e-01, 3.10355639e-02, 1.53539405e-01],
+           [1.62522770e-01, 3.12091760e-02, 1.60094125e-01],
+           [1.67631454e-01, 3.12849946e-02, 1.66767503e-01],
+           [1.72718985e-01, 3.12602060e-02, 1.73562572e-01],
+           [1.77783306e-01, 3.11317439e-02, 1.80482912e-01],
[…]
