+++ .codecov.yml
@@ -1,3 +1,7 @@
+codecov:
+    notify:
+        require_ci_to_pass: yes
+
 coverage:
     round: nearest
     precision: 2

[…]
