### 2020-01-31 a7a9fc2
Improved the documentation a bit on how to import custom colormaps.
