### 2020-01-31 #2004 @tacaswell
As a side note, @1313e proposed some of these color maps for inclusion in Matplotlib and we suggested a stand-alone package.  I think that this has ended up in a vastly better state than if the color maps had been added to core.  There is great value in an curated set of color maps that are nicely name spaced and more opinionated than we can be on core.
