### 2020-02-01 #2004 @wafels
I approve this paper and look forward to seeing the colormaps used in the scientific community.
