### 2020-02-03 d18b750
Added the 'apple' colormap. Hopefully fixed the placement of Figure 2 in the JOSS paper.
