+++ cmasher/colormaps/cm_apple.txt
@@ -1,0 +1,256 @@
+0.000000000000000000e+00 0.000000000000000000e+00 0.000000000000000000e+00
+3.452201729999999913e-04 1.919086299999999963e-04 2.222953309999999905e-04
+1.253166250000000087e-03 6.451711900000000471e-04 7.665588770000000482e-04
+2.703462279999999975e-03 1.297985480000000014e-03 1.579315970000000000e-03
+4.702119319999999852e-03 2.118274749999999890e-03 2.635262209999999789e-03
+7.261134179999999966e-03 3.083372940000000170e-03 3.917310789999999682e-03
+1.039480800000000017e-02 4.175484479999999957e-03 5.412344020000000276e-03
+1.411863749999999959e-02 5.379761989999999779e-03 7.109516280000000368e-03
+1.844900340000000041e-02 6.683264310000000021e-03 8.999391810000000047e-03
+2.340313509999999986e-02 8.074314560000000335e-03 1.107343910000000049e-02
+2.899795130000000151e-02 9.542528839999999887e-03 1.332384160000000053e-02
+3.525112670000000309e-02 1.107815000000000015e-02 1.574318010000000032e-02
+4.213124859999999722e-02 1.267185369999999922e-02 1.832427559999999858e-02
+4.909383159999999857e-02 1.431527439999999950e-02 2.106029790000000027e-02
+5.602828779999999981e-02 1.599998609999999863e-02 2.394438939999999835e-02
+6.294068969999999652e-02 1.771813590000000105e-02 2.696983500000000103e-02
+6.983598219999999357e-02 1.946225479999999855e-02 3.012997900000000118e-02
+7.671927900000000133e-02 2.122460779999999977e-02 3.341795499999999919e-02
+8.359351830000000261e-02 2.299852889999999886e-02 3.682717690000000127e-02
+9.046311460000000526e-02 2.477635140000000041e-02 4.035051399999999705e-02
+9.732985489999999851e-02 2.655195119999999992e-02 4.383742570000000172e-02
+1.041973830000000045e-01 2.831786509999999896e-02 4.724804929999999736e-02
+1.110670770000000002e-01 3.006813919999999987e-02 5.058590360000000258e-02
+1.179415230000000064e-01 3.179584900000000103e-02 5.385217369999999865e-02
+1.248224039999999979e-01 3.349468290000000292e-02 5.704795499999999742e-02
+1.317108909999999966e-01 3.515874110000000163e-02 6.017419999999999713e-02
+1.386094549999999925e-01 3.678085329999999931e-02 6.323111319999999425e-02
+1.455185689999999865e-01 3.835564980000000235e-02 6.621933119999999506e-02
+1.524394030000000011e-01 3.987710280000000135e-02 6.913901159999999435e-02
+1.593735470000000043e-01 4.131817890000000215e-02 7.198990880000000203e-02
+1.663218509999999928e-01 4.266018249999999762e-02 7.477183980000000341e-02
+1.732848989999999922e-01 4.391024550000000026e-02 7.748453630000000369e-02
+1.802635290000000057e-01 4.506769289999999761e-02 8.012743210000000671e-02
+1.872584890000000002e-01 4.613165959999999954e-02 8.269980389999999348e-02
+1.942704460000000133e-01 4.710109540000000095e-02 8.520077530000000343e-02
+2.012999819999999884e-01 4.797476779999999830e-02 8.762931969999999959e-02
+2.083475999999999939e-01 4.875126429999999955e-02 8.998426210000000369e-02
+2.154137190000000035e-01 4.942899409999999827e-02 9.226428070000000647e-02
+2.224986750000000069e-01 5.000618889999999817e-02 9.446790630000000133e-02
[…]
