### 2020-02-03 cfba1e8
Added some logic to when Travis deploys a new version on PyPI.
