### 2020-02-03 #2004 @1313e
@mbobra I have just fixed the placement of Figure 2 in the paper.
Shall I make a GitHub release for v1.2.2, which automatically makes a Zenodo archive, and post the DOI here?
