### 2020-02-04 #2004 @mbobra
> @mbobra I have just fixed the placement of Figure 2 in the paper.
> Shall I make a GitHub release for v1.2.2, which automatically makes a Zenodo archive, and post the DOI here?

Yes please!
