### 2020-02-04 #2004 @1313e
Also, the version is v1.2.2 now.
Not sure if that needs to be changed.
