### 2020-02-05 #2004 @kthyng
@1313e Please take a look at references Sharpe et al, Smith et al 2018, and Smith et al 2019. For each of these either the spacing is odd or the author list seems incomplete. Please verify the best way to reference the Smith packages.
