### 2020-02-05 #2004 @1313e
> @1313e Please take a look at references Sharpe et al, Smith et al 2018, and Smith et al 2019. For each of these either the spacing is odd or the author list seems incomplete. Please verify the best way to reference the Smith packages.

Could you elaborate on what is wrong with those references?
Because, I cannot see anything weird or wrong with any of the references I'm using.

> Hi @1313e! I am taking over from here. Can you update the metadata in your zenodo archive to match your paper title and authors?

Of course. Forgot that automated Zenodo archiving will obviously not set those things.
