### 2020-02-06 #2004 @kyleniemeyer
@1313e if it helps, it's possible to add a `.zenodo.json` file to a repository which will specify these metadata, though I can't find an official guide on how to do this/what to include... here is an example though: https://github.com/cboettig/nonparametric-bayes/blob/master/.zenodo.json
