### 2020-02-06 5cea6ec
Edits to bib file
