+++ joss_paper/paper.bib
@@ -190,7 +190,7 @@ abstract = {Literature that describes the prevalence of inherited red-green colo
 }
 
 @inproceedings{Sharpe99,
-  title={1 Opsin genes , cone photopigments , color vision , and color blindness},
+  title={Opsin genes, cone photopigments, color vision, and color blindness},
   booktitle={Color Vision: From Genes to Perception},
   publisher={Cambridge University Press},
   pages={3--51},
