### 2020-02-06 #2004 @kthyng
@1313e Ok, I edited the Sharpe entry directly then. My changes are in PR [#4](https://github.com/1313e/CMasher/pull/4) and you can merge them if you agree. I looked up the details for Nathaniel's papers and they look accurate.
