### 2020-02-06 #2004 @kthyng
By the way, since you're interested in colormaps, you might be interested in my paper on them!
http://tos.org/oceanography/assets/docs/29-3_thyng.pdf
