### 2020-02-06 #2004 @1313e
> @1313e Ok, I edited the Sharpe entry directly then. My changes are in PR [#4](https://github.com/1313e/CMasher/pull/4) and you can merge them if you agree. I looked up the details for Nathaniel's papers and they look accurate.

Oooh, those spacings.
Kinda weird that they are there, as they don't appear in my bibtex database.

> By the way, since you're interested in colormaps, you might be interested in my paper on them!
> http://tos.org/oceanography/assets/docs/29-3_thyng.pdf

I actually know that paper already and somehow completely forgot to add a citation to `cmocean` in the paper.
Want me to add a citation to that paper?
