### 2020-02-06 #2004 @kthyng
> Want me to add a citation to that paper?

I'm not reviewing your paper — I just thought you'd be interested in work in the same area.

I'll proceed with the final steps
