### 2020-02-07 #2004 @1313e
@kthyng I am not entirely sure why, but when I go to the paper [here](https://joss.theoj.org/papers/10.21105/joss.02004), the citation information and the BibTeX entry are kinda messed up (wrong year, volume and issue).
I initially thought that this is maybe caused by the paper just being published, but the other just-published papers do not seem to have that problem.
Any ideas what this is?
