### 2020-02-08 #2004 @mbobra
It also looks like [the tweet](https://twitter.com/JOSS_TheOJ/status/1225459954410192897) has the wrong citation information.
