### 2020-02-11 #2004 @arfon
> It also looks like the tweet has the wrong citation information.

Still? The information probably took a little while to update on Twitter's side but looks good to me now. Can you check again @mbobra?
