### 2020-02-11 #2004 @1313e
> I am just clicking on the link above that says "Tweet for this paper". It still says "1(1), 200..." which cannot be right . Edit: I am clicking from my laptop (the link may look different on mobile).

The tweet does not contain any citation info?
