### 2020-02-11 #2004 @mbobra
This is what I am seeing:
![citation](https://user-images.githubusercontent.com/8431057/74289812-dfd8c000-4ce4-11ea-91c9-e3968b591344.jpg)
