### 2020-02-12 #2004 @arfon
Yeah, I think this is cached. The only thing we can do here is delete the original tweet an resend it from the JOSS account (which I'm more than happy to).

Lemme know what you'd like us to do @1313e.
