### 2020-02-12 #2004 @1313e
As I doubt anyone will use citation information from a preview, don't bother.
