### 2020-03-10 #2004 @1313e
@arfon Do you, by any chance, know how long it usually takes for ADS to index new JOSS papers?
I was looking at ADS, and it seems that it indexes new JOSS papers that it flags as 'astronomy-related' incredibly quickly, whereas other JOSS papers take quite some time.
(All JOSS papers published this year that are indexed by ADS, which is 5 in total, are flagged as 'astronomy-related', which I doubt is a coincidence.)
