### 2020-03-10 #2004 @mbobra
@1313e If you want to speed up the process, you can ask ADS to index your paper [using this form](http://adsabs.harvard.edu/adsfeedback/submit_abstract.php) and they will likely approve it right away.
